EMBR-7455 chore: move type to common folder, revert 'manual' change
