+++ src/common/index.ts
@@ -1,0 +1,1 @@
+export type { VisibilityStateDocument } from './types.js';
