+++ src/common/types.ts
@@ -1,0 +1,4 @@
+// Useful for testing so that we can pass in a document-like object and change its visibilityState
+export interface VisibilityStateDocument {
+  visibilityState: DocumentVisibilityState;
+}
