+++ src/instrumentations/session/SpanSessionVisibilityInstrumentation/types.ts
@@ -1,9 +1,5 @@
 import type { EmbraceInstrumentationBaseArgs } from '../../EmbraceInstrumentationBase/index.js';
-
-// Useful for testing so that we can pass in a document-like object and change its visibilityState
-export interface VisibilityStateDocument {
-  visibilityState: DocumentVisibilityState;
-}
+import type { VisibilityStateDocument } from '../../../common/index.js';
 
 export type SpanSessionVisibilityInstrumentationArgs = Pick<
   EmbraceInstrumentationBaseArgs,
