+++ src/managers/EmbraceSpanSessionManager/types.ts
@@ -1,10 +1,6 @@
 import type { DiagLogger } from '@opentelemetry/api';
 import type { PerformanceManager } from '../../utils/index.js';
-
-// Useful for testing so that we can pass in a document-like object and change its visibilityState
-export interface VisibilityStateDocument {
-  visibilityState: DocumentVisibilityState;
-}
+import type { VisibilityStateDocument } from '../../common/index.js';
 
 export interface EmbraceSpanSessionManagerArgs {
   diag?: DiagLogger;
